+++ Modules/SharePointDsc/Examples/Resources/SPWebAppPropertyBag/1-AddWebAppPropertyBag.ps1
@@ -1,0 +1,27 @@
+<#
+.EXAMPLE
+    This example shows how add property bag value in a web application.
+#>
+
+Configuration Example
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [PSCredential]
+        $SetupAccount
+    )
+
+    Import-DscResource -ModuleName SharePointDsc
+
+    node localhost
+    {
+        SPWebAppPropertyBag APPLICATION_APPCodeProperty
+        {
+            PsDscRunAsCredential = $SetupAccount
+            Key = "KeyToAdd"
+            Value = "ValueToAddOrModify"
+            Ensure = "Present"
+        }
+    }
+}