+++ Modules/SharePointDsc/Examples/Resources/SPWebAppPropertyBag/1-RemoveWebAppPropertyBag.ps1
@@ -1,0 +1,26 @@
+<#
+.EXAMPLE
+    This example shows how remove a property bag value in a web application.
+#>
+
+Configuration Example
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [PSCredential]
+        $SetupAccount
+    )
+
+    Import-DscResource -ModuleName SharePointDsc
+
+    node localhost
+    {
+        SPWebAppPropertyBag APPLICATION_APPCodeProperty
+        {
+            PsDscRunAsCredential = $SetupAccount
+            Key = "KeyToRemove"
+            Ensure = "Absent"
+        }
+    }
+}